[Feat] 스크랩 페이지 로직 분리
